Develop

+++ walking-beans-backend/src/main/java/walking_beans/walking_beans_backend/model/dto/Payments.java
@@ -13,9 +13,12 @@ import lombok.*;
 public class Payments {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int paymentId;
-    private int orderId;
+    private long paymentId;
+
+    private long orderId;
+
     private String paymentMethod;
+
     private String paymentStatus;
 
 }
